test/setpython: give meaningful feedback even with no Python redirect
When invoked with no arguments, 'setpython' reports the path to which
the 'python' command is currently redirected. However, if not currently
redirected, it emits an unhelpful error message:

    ls: python: No such file or directory

Fix this case so that it instead reports the path of the system default
Python.

While at it, also make the normal output more terse; rather than showing
the output of "ls -l", just report the path of the chosen Python.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>