test/Makefile: work around ancient MacOS GNU make shortcoming
GNU make on modern stock MacOS (v10.12.3) is stuck at version 3.81
which was released on 2006-04-01, nearly 11 years ago. It has the
quirky behavior that the 'setpython' script in the 'test' directory
is not found, despite the 'test' directory being in PATH. Given:

    default:
        setpython python2

The invocation fails with:

    make: setpython: No such file or directory

Minor variations of the command side-step the failure. For instance,
any of the following work fine:

    default:
        ./setpython python2
        true && setpython python2
        if true; then setpython python2; fi

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>