test/setpython: use bare 'python' if it matches requested version
The default Makefile target attempts running all tests under Python 2
and Python 3 using commands 'python2' and 'python3', respectively. If
these exact command names are not found on the system, then the tests
are not run at all, even if a Python interpreter of the requested
version is available named simply 'python'.

Fix this unfortunate shortcoming by also checking the bare name 'python'
(if installed) for the requested Python version and using it if it
matches.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>