test/Makefile: provide feedback by announcing long-running tests
Many tests take a long time to run, and most of them provide some sort
of feedback to let the user know that something is happening.

The 'pylint' test, however, provides no feedback whatsoever and takes
quite a long time to run. Moreover, it is the very first test run by
the default Makefile target, with the result that the user is met by
a very lengthy silence immediately upon invoking 'make'. Address this
by announcing that the 'pylint' test is running.

While at it, also announce the built-in reposurgeon test.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>