test/Makefile: skip tests for which prerequisites are missing
The Makefile already robustly skips specific tests for which
prerequisite commands 'git', 'hg', 'svn', 'cvs', and 'cvs-fast-export'
are missing, which allows for at least partial testing on systems not
fully provisioned.

However, it is not at all forgiving when 'pylint', 'python2', or
'python3' are not installed and aborts all testing immediately upon
finding one of these commands missing. Be consistent and instead
robustly skip tests specific to those commands if not installed.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>