reposurgeon: create selection parsing class
The RepoSurgeon event selection parser and evaluator is powerful and
expressive, and may have applications outside of event selection,
however, it is coupled tightly with events, thus it is not presently
possible to employ it elsewhere.

In order to make it generally useful the event-specific logic needs to
be factored out. As a first step, introduce a general-purpose selection
parsing class and move low-level utility routines there verbatim from
the RepoSurgeon class.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>