eval_polyrange: generalize detection of polyrange initials
Much of the functionality of eval_polyrange() is generic, not specific
to RepoSurgeon events. For instance, its handling of numbers, "$",
comma, and ranges may be useful outside of events. The polyrange
"initials" set, however, has event-specific characters (":", "<") baked
into it. To make it re-usable, factor out the common initials and
provide a way for subclasses to augment the set of initials.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>