eval_polyrange: make span-resolver responsible for reporting open ".."
The goal is to rid eval_polyrange() of RepoSurgeon event-specific
knowledge so that it can be relocated to the generic SelectionParser.
In aid of this, the various "recognizers" (number, "$", comma, span,
etc.) will be factored out of eval_polyrange into overridable method
eval_atom(). This new method, however, will not have access to the
'selection' array being built by eval_polyrange(), thus it won't be able
to detect a missing start of span due to lack of knowledge of what came
before it (if anything). Therefore, make detection and reporting of
missing start of span the responsibility of poly_range()'s
span-resolver, which does have access to the full selection[] array.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>