eval_atom: override to handle event-specific logic specially
The goal is to rid eval_polyrange() of RepoSurgeon event-specific
knowledge so that it can be relocated to the generic selection parser.
Toward this end, factor out the remaining event-specific logic into a
RepoSurgeon override of eval_atom().

This is a purely textual relocation with the minor exception of a small
amount of boilerplate at the start and end of the eval_atom() override,
plus one retired 'continue' statement since the code now resides in a
stand-alone function rather than within a 'while' loop.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>