eval_polyrange: generalize client-specific possible-polyrange detection
The goal is to rid eval_polyrange() of RepoSurgeon event-specific
knowledge so that it can be relocated to the generic SelectionParser
class. eval_polyrange() currently contains hard-coded logic to avoid
confusing an event-specific polyrange "<...>" selector with '<' input
redirection. Generalize detection of a possible polyrange and factor out
this special case into the RepoSurgeon subclass of SelectionParser.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>