eval_textsearch: factor out RepoSurgeon event-specific searching
The goal is to rid eval_textsearch() of RepoSurgeon event-specific
knowledge so that it can be relocated to the generic SelectionParser.
Toward this end, move the event-specific query modifiers and search
logic out of eval_textsearch() and into new method textsearch_items().

This is a purely textual relocation (plus one level outdent).