eval_visibility: generalize '=X' type-letter recognition
The '=X' visibility set concept is not specific to RepoSurgeon events,
thus ought to be handled by the generic selection parser, however,
eval_visibility() has event-specific knowledge baked into it. As a
first step toward generalizing this functionality, extract the mapping
of type-letter-to-lambda table to a method which subclasses of the
generic parser can override.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>