debug_lexer: generalize for compile/evaluate paradigm
SelectionParser conflates parsing and evaluation of a selection
expression into a single operation. The goal is to separate the two so
that an expression can be compiled once and evaluated multiple times.

As a first step toward this goal, upgrade @debug_lexer to function
properly during either phase. In particular, 'self.line' will be
available only during expression compilation (parsing), and
'preselection' will be available only during evaluation.

An alternative approach would be to have different decorators for
parsing and evaluation (say, @debug_parser and @debug_eval), however,
the basic functionality of each would be so similar, and they would
duplicate or share so much code, that the benefit of having two
decorators rather than the single (somewhat simpler) @debug_lexer does
not currently seem warranted.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>