eval_term: factor out generic parsing logic into SelectionParser
With the exception of invoking RepoSurgeon-specific eval_pathset(), all
the logic in eval_term(), such as set negation "~" and expression
grouping "(...)", is generic, thus can sensibly reside in a generic
selection parser. Therefore, relocate it to SelectionParser.eval_term()
and let RepoSurgeon.eval_term() override it to handle the eval_pathset()
special case.

This is purely a textual relocation of the generic portion of
eval_term().

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>