eval_disjunct: convert to compile/evaluate paradigm
SelectionParser conflates parsing and evaluation of a selection
expression into a single operation. The goal is to separate the two so
that an expression can be compiled once and evaluated multiple times.

As a step toward this goal, convert eval_disjunct() to the
compile/evaluate paradigm by separating parse-time logic from
evaluation-time logic.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>