set_selection_set: clarify intent of selection-clearing conditional
The final conditional of set_selection_set() checks if parsing actually
consumed any of the input line, however, it is somewhat difficult to
grasp the intention due to the similarities of variable names involved
('self.line' which was parsed, and 'line', the original input). Make
the intent clearer by naming the original input 'orig_line'.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>