set_selection_set: clarify intent of named-lookup transformation
It isn't immediately clear to the newcomer why the "name" to "<name>"
transformation for a named lookup is applied to the line to be parsed
('self.line') but not to the input 'line'. In practice, it doesn't
matter, since the input line is only ever consulted thereafter to see if
it differs from the parsed line, which means that the parser consumed
some portion of the input. This simple check works correctly whether the
the "name" to "<name>" transformation is applied to the original input
or to the line to be parsed. However, it is semantically incorrect
since the parsed line is expected to differ from the input line in some
predictable fashion (for instance, the line after parsing will be a
subset of the input line). Applying the transformation only to the line
to be parsed breaks that assumption. As an aid to future readers, avoid
this semantic ambiguity by performing the transformation on the original
input instead and letting the line to be parsed inherit it.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>