eval_expression: partially convert to compile/evaluate paradigm
SelectionParser conflates parsing and evaluation of a selection
expression into a single operation. The goal is to separate the two so
that an expression can be compiled once and evaluated multiple times.

Rather than converting SelectionParser to the new paradigm in its
entirety, the conversion will be done piecemeal. The simplest way to do
is bottom-up, however, eval_expression() is invoked recursively, so,
even though it is the top-most parser production, in a sense, it is also
the bottom-most.

To work around this paradox, partially convert eval_expression() to the
compile/evaluate paradigm by changing its signature to return a
"machine" capable of evaluating the selection expression, rather than
directly returning the result of evaluation. A (much) later change will
finalize the conversion.

Since recursive callers of eval_expression() have not yet been converted
to the compile/evaluate paradigm, they presently need to immediately
"run" the machine returned by eval_expression(). This is a temporary
measure; eventually, the entire machine will be run en-bloc by the
top-level caller of parse_expression(). Temporary code is commented as
"TEMPORARY" and all such code will be removed by the end of the
conversion.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>