parse_polyrange: add compile-time sanity checks
Detect obvious span malformations, such as "..3" or "3.." at compile
time.

Note: Evaluation-time checks for open-ended spans are still needed, thus
are not removed, since a valid compile-time expression may nevertheless
become an open-ended span when evaluated. For instance, in an expression
such as "3../foo/", /foo/ may evaluate to an empty selection, which
effectively makes the expression "3..".

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>