eval_funcall: improve debugging output
Now that selection parsing and evaluation are separate, it can be
difficult to associate the name of the @func() function -- which is
only shown at parse time -- with the result of its application shown
much later at evaluation time. Improve the situation by showing
evaluation-time entry and exit to the @func() handler itself, which
makes it quite obvious what transformation is being applied to the
input.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>