parse_expression: finalize conversion to compile/evaluate paradigm
Finalize conversion of parse_expression() to the compile/evaluate
paradigm.

This removes the last of the temporary "run machine immediately" code
(commented as "TEMPORARY") which allowed the conversion to proceed
piecemeal rather than being done as one giant step. The only remaining
code which "runs" the machine returned by parse_expression() is the
top-level caller, set_selection_set(), and the returned machine itself
is now complete (the result of the full parse process).

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>