attribution: prepend/append: infer date and name or email
As a convenience when inserting a new attribution, make the date and
name or email optional and infer the missing values, if possible, from
existing attributions. For instance, if the name is omitted but the
provided email address matches the email of an existing attribution,
take the name from that attribution. Preference is given to the
attribution before or after which the new attribution is being inserted.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>