attribution: add 'prepend' and 'append' actions
Add commands "attribution [<selection>] prepend" and "attribution
[<selection>] append" which insert a new attribution. 'prepend' inserts
a new attribution before the first attribution selected; 'append'
inserts after the last selected. The new attribution has the same type
(committer, author, tagger) as the one before or after which it is being
inserted.

As a convenience, if the selection is empty or not specified, a new
author attribution is inserted before the first existing author
attribution for 'prepend', or after the last author for 'append'.

It is an error to insert a committer or tagger attribution (though this
restriction may be lifted if commits and tags are someday upgraded to
support more than one committer or tagger, respectively).

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>