attribution: add 'resolve' action
Add command "attribution <selection> resolve" which merely resolves the
attribution selection expression and echos the resulting attribution
number set. The remainder of the line after the command is used as a
label for the output.

Implemented mainly for regression testing; complements the top-level
'resolve' command which resolves an event selection expression and echos
the result.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>