reorder_commits: perform rudimentary fileops validation
Re-ordering commits can lead to fileops inconsistencies. For instance,
a re-ordered commit may delete, rename, or copy a file before it has
ever been created. Likewise, a commit renaming a file may be placed
ahead of commits which still reference the file by the original name.

Add rudimentary fileops validation to warn about problems discovered and
remove or fix broken fileops. The types of inconsistencies detected by
the current implementation is relatively limited, and it only checks for
inconsistencies in commits in the specified range plus immediate
children. One or both limitations may be addressed in the future.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>