Implement --preserve, *not* as the default.
This is a significant change in behavior. Now, empty Subversion branch
notes normally turn into bare resets rather than being tagified. Their
comment text is shipped to stdout in a message.

As presently is, the code onteracts badly with the handling of tip deletes;
You can see this is the branchreplacee tear, where three resetare generated but
one is clobbered by the tipdelete handling before being emitted.

Rewriting that code is the next thing on the list.