selectionparser: fix infinite recursion when chaining |, &, ? operators
Selection operators '|' (disjunction), '&' (conjunction), and '?'
(neighbors) can be chained (e.g. "1|3|5", "1&3&5", "3??"). Parsing of
each operator happens in a loop. For instance:

    def parse_disjunct(self):
        op = self.parse_conjunct()
        while True:
            if self.peek() != '|':
                break
            self.pop()
            op2 = self.parse_conjunct()
            op1 = op
            op = lambda p: self.eval_disjunct(p, op1, op2)
        return op

The goal is to build an "evaluation machine" of the chained elements.
For example, given "1|3|5" the resulting evaluation machine should be:

    disjunct(disjunct(1,3),5)

Unfortunately, due to Python's late binding of variables in lambdas, all
lambdas in the constructed machine actually see only the final values of
'op1' and 'op2', thus the machine looks like this:

    disjunct(disjunct(disjunct(...,5),5),5)

where '...' represents infinite recursion of the 'disjunct' evaluator.

The problem of late binding was understood when the parsers for these
operators were implemented, and the author took defensive measure by
ensuring that a variable defined outside the loop was assigned to a new
variable inside the loop before using it in the lambda. Unfortunately,
the author (a casual Python programmer) forgot or did not know "scoping
rules" within Python loops and thought that variables assigned within
the loop, but never seen outside it, would be created anew upon each
iteration, thus were safe to reference in the lambda. However, this is
not how Python works, so this tactic failed.

Fix the problem by binding the values immediately within the lambda's
closure.

This problem was introduced when the selection parser was converted to a
"compile / evaluate" paradigm; it originally evaluated as it parsed the
selection expression.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>