repotool: fix incorrect git-branch invocation
To obtain a branch list from Git, repotool invokes "git branch -l",
however, "-l" is short for --create-reflog, not for --list. The "-l" or
--create-reflog option is meant to be used only when creating a new
branch (e.g. "git branch -l newbranch"), and it is an accident of
implementation (within Git) that "git branch -l" silently ignores the
"-l" and lists branches as if bare "git branch" was used.

Upcoming versions of Git deprecate "-l" as shorthand for --create-reflog
and emit a suitable deprecation notice. Aside from being unsightly for
normal use of 'repotool', the deprecation notice also causes tests to
fail which expect 'repotool' to be silent.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>