setpypy: fix case when bare 'python' matches requested Python version
As originally implemented, 'setpython' (before being renamed to
'setpypy') required a Python executable to exist matching the exact name
of the requested Python version. For instance, when "python2" was
requested, a "python2" executable had to be installed on the system for
'setpython' to make that Python version the active one, even if a
bare-named "python" happened to match the desired version.

<2017-03-09T14:56:50-05:00!sunshine@sunshineco.com> updated 'setpython'
to also accept the bare-named "python" if it matched the desired
version. However, this particular case was broken by
<2018-01-02T00:48:16-05:00!esr@thyrsus.com> which made 'pypy' the
default interpreter, rather than 'python', by changing the shebang lines
of Python scripts from "#!/usr/bin/env python" to "#!/usr/bin/env pypy"
(and renaming 'setpython' to 'setpypy').

When 'setpython' noticed that the bare-named "python" matched the
desired Python version, it simply removed the local symbolic link named
'python', allowing "#!/usr/bin/env python" to find the installed
bare-named "python", as usual. The change to "#!/usr/bin/env pypy",
broke this because asking /usr/bin/env to run "pypy" will never result
in it running the bare-named "python". To fix this, we need to actively
create a local 'pypy' symbolic link pointing at the bare-named "python".

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>