setpypy: drop now-unneeded Python version check workaround
Prior to <2018-01-02T00:48:16-05:00!esr@thyrsus.com> which defaulted the
project to 'pypy' instead of 'python', and which renamed 'setpython' to
'setpypy', a workaround was needed to avoid the local 'python' symbolic
link from fooling the version check of whatever bare-named "python" was
installed on the system. The workaround was added by
<2017-03-12T12:50:01-04:00!sunshine@sunshineco.com>.

However, this workaround is no longer needed since the local 'pypy'
symbolic link will never accidentally be invoked when checking the
Python version via "python -V", therefore, remove it.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>