setpypy: work around MacOS refusal to execute symbolic link in PATH
'setpypy' operates by creating a symbolic link to the desired Python
interpreter (whether it be "pypy", "python2", or "python3"), and
test/Makefile adds the local "test/" directory to PATH. This allows the
reposurgeon suite to be tested under various Python implementations
since the shebang line in each program is "#!/usr/bin/env pypy", which
will find the local 'pypy' symbolic link, if present, or the installed
'pypy' command, if not.

Unfortunately, MacOS outright refuses to consider a symbolic link found
via PATH to be a valid executable worthy of running, which renders this
schemes unusable. Work around the problem by having 'setpypy' create a
one-liner shell script named 'pypy' (instead of a symbolic link), which
MacOS is happy to invoke.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>