test: fall back to 'python' if 'pypy' not installed
Rather than assuming unconditionally that 'pypy' is installed when
running the tests, fall back to running the tests under plain 'python'
if 'pypy' is unavailable. Skip the tests only if neither is installed.

While at it, fix invocation of the "built-in" tests so they too fall
back to 'python' if 'pypy' is unavailable, and are cleanly skipped if
neither is.

Finally, since "setpypy python" may create a local 'test/pypy' under
this new arrangement, clean it up upon completion, as is already done
for "make portcheck".

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>