SelectionParser: don't expect "$" to be resolved at parse time
SelectionParser has distinct parse and evaluate phases. The number of
items available for selection is known only at evaluation time, so it is
impossible to resolve "$" at parse time, yet the Python implementation
incorrectly does so. That "$" actually works in Python, despite this
bug, is pure accident due to how Python interprets negative array
indices relative to the end of the array. So, when parse_atom()
evaluates len(allitems)-1, the result is unconditionally -1 (since
'allitems' is not assigned at parse time), and when eval_polyrange()
evaluates allitems[-1] it ends up retrieving the last item of the array,
as expected (even though by accident). Go, however, does not support
negative array indices, so such accidentally-correct behavior is
impossible. To correctly deal with "$", parseAtom() must instead
indicate that "$" was seen in the input at that position, and leave it
up to evalPolyrange() to resolve it.