hasReference: don't copy sync.Mutex embedded in regexp.Regexp
A "for x := range" loop places a copy of the value in 'x'. In this
case, the value is a regexp.Regexp, which embeds a sync.Mutex.
Although not harmful in this case, copying a sync.Mutex is typically
avoided, and is unnecessarily wasteful. Avoid copying by instead
referencing the Regexp directly. (Reported by "go vet".)