implement the misc help functions
HelpFunctions turned out to be more interesting than I expected. With
no real docstrings for functions, we have to fall back to a map of
them. I decided to still use the reflect package to get the full list
of methods, so that new ones will show up in the help even if we
forget to document them.