SelectionParser: implement evalVisiblity()
This implementation differs slightly from the Python version but should
be functionally identical. In particular, the Python version
unnecessarily resolves "allitems[i]" even though 'i' is _always_ the
same value as that stored at "allitems[i]". This "oops" was an oversight
from when the SelectionParser was converted to a two-step parse/evaluate
paradigm.  That unnecessary complexity is dropped in the Go version.