SelectionParser: implement parseFuncall()
Given @foo(), the Python implementation uses reflection to find a
handler function named foo_handler(). Although this is convenient in
Python, it is not so in Go. Instead, we'll employ a map associating
selection-parser function name to the actual handler function.
Subclasses of SelectionParser can implement a functions() method to
augment the base set of functions.