SelectionParser: support subclass polymorphism
Subclasses of SelectionParser overload some of its methods to extend
its behavior. Other methods in SelectionParser need to call those
overloaded methods, however, Go doesn't support object-oriented
programming, so this sort of automatic polymorphism doesn't exist in
Go. Although it may make sense in the future to re-work this
architecture to be more Go-idiomatic, the present translation from
Python to Go is intended to be as straightforward literal conversion
as possible. As such, introduce manual subclass polymorphism support
by allowing subclasses to register themselves with the base class
(which they will embed) as "implementation" (or "delegate").