fix SelectionParser.parseAtom() not anchoring number matching
By neglecting to anchor its recognition of numbers to the start of
input, parseAtom() incorrectly classifies input as a number even when
it's not. For instance, in the expression, "1..4", the first call to
parseAtom() correctly identifies the number "1" and strips it, returning
"..4". But, because number recognition is not anchored to the start of
input, the next call incorrectly thinks "..4" is also a number ("4"),
strips a "." from the start of string, returning ".4".  Ditto for the
third call, finding "4" and stripping ".", to return "4".  The final
call to parseAtom() correctly identifies "4" and returns it. Thus, input
"1..4" results in the return of four numbers (1, 4, 4, 4) which
collapses to incorrect set [1, 4].

Fix this by anchoring number recognition to start of input, so the
result of "1..4" is three tokens (1, "..", 4) which is correctly
expanded to set [1, 2, 3, 4].