add helpers for parsing and evaluating selection sets
The cleaner calling convention is nice, and it will keep us from
forgetting about extra steps like the isNamed stuff.

Use setSelectionSet if you want to unconditionally evaluate the set
against the "chosen" repository. Use parse+eval if it's conditional or
you want to control which repository to evaluate against.