allow the user to define and run macros
Uses some small goroutines to run an inner command loop for both input
of macros and playback of macros. The outer loop waits for the
goroutine to send back a signal indicating completion; from
Kommandant's perspective, the control flow is perfectly ordinary.

Compare this to the current implementation of DoScript, which rewrites
most of Kommandant's CmdLoop method. I'll rewrite it to use the same
goroutine trick soon enough.