the script command needs to help out just in case the script wants to define a macro
This would be so much simpler if scripts didn't also have to have
heredocs and positional variables. Alternatively, we could enable
heredocs and positional variables for the interactive console as
well. Either way, this would simplify down to simply setting the
rs.cmd.Stdin to the file and setting up a goroutine to call
cmd.CmdLoop just like the DoDo and DoDefine methods.