add fastOrderedIntSet to complement orderedIntSet
fastOrderedIntSet is like orderedIntSet but optimizes for speed at the
expense of space (whereas, orderedIntSet optimizes for space at the
expense of speed). fastOrderedIntSet is appropriate for applications in
which operations such as intersection or union may be applied to very
large sets, such as when operating on selections via SelectionParser.