normalize formatting of all Go code via 'gofmt'
Take advantage of this quiescent time (when the project is not being
heavily edited by multiple developers) to normalize the formatting of
all the Go code.

This should ease the life of developers whose editors are configured
to run 'gofmt' automatically upon "Save" by eliminating the need to
separate out their actual changes (when it's time to commit them to Git)
from the "noise" caused by 'gofmt' modifying code they are not otherwise
touching.