SelectionParser: drop undesirable method stub
The base SelectionParser class does not itself understand any
"visibility typeletters". Subclasses of SelectionParser, such as
Reposurgeon and AttributionEditor.SelParser, are expected to
implement visibilityTypeletters() if they need to support the
feature. parseVisibility() and evalVibisility() explicitly check if
visibilityTypeletters() is implemented, and gracefully handle the case
when it is not.

When visibilityTypeletters() is implemented, it is expected to return a
meaningful mapping between typeletters and predicate lambdas;
evalVibisility() is not prepared for visibilityTypeletters() returning
nil instead of a valid mapping. Therefore, remove the unwanted and
unneeded SelectionParser.visibilityTypeletters() fallback which merely
returns nil.