evalVisibility: fix crash due to bogus slice population
The []predicates slice is intended to hold a predicate lambda
corresponding to each specified visibility typeletter. For instance, the
expression "=CT" should populate it as [<commit-lambda>,<tag-lambda>].
However, due to a bug, it mispopulates the slice with nil where lambdas
are expected, for instance [nil,nil,<commit-lambda>,<tag-lambda>], which
leads to a crash when it tries invoking nil. Fix this.