reposurgeon: stop adding extra newline to announce() messages
announce() automatically terminates the message with a newline, so there
is no need to do so manually. Moreover, the extra newline makes the
output of these commands ("resolve", "verbose", "echo") differ from the
Python version. This can be problematic for test cases which often
capture literal command output and compare it to "expected" output.