newLineParse: stop adding cryptic prefix to error messages
Messages from recoverable errors during line parsing are meant to be
user-facing, so they should be easy to understand and reflective of what
the user was attempting. Prefixing them with a cryptic implementation
detail doesn't help, so drop the prefix.

Moreover, the prefix makes the output differ from the Python version,
which can be problematic for test cases which often capture literal
command output and compare it to "expected" output.