newLineParse: drop unnecessary "inner function" indirection
Although returning an error from a function is a typical Go idiom, there
is little point in making newLineParseInner() do so when its only
caller, newLineParse(), immediately and unconditionally throws that
error as an exception. The extra indirection merely serves to complicate
the code unnecessarily, so fold newLineParse() and newLineParseInner()
together and just raise exceptions like most other code in
reposurgeon.go.