DoResolve: stop prefixing "resolve" output with unwanted noise
The sole purpose of the "resolve" command is to generate very specific
literal output. It's output is not meant to be some sort of
informational message leting the user know what the command is doing.
As such, use of announce() to generate its output is semantically
incorrect. Moreover, announce() prefixes the messages with unwanted
noise, making the output differ from the Python version, which can be
problematic for test cases which often capture literal command output
and compare it to "expected" output.