getAttr: fix crash when extracting field from object reference
getAttr() only works when handed an actual object; it crashes when given
a pointer to an object. This makes it less useful than it should be
since a common use-case is to extract a field value from a generic
Event, which is always a reference (*Commit, *Tag, etc.) to some
concrete type. Fix getAttr() to be able to extract field values from
both actual objects and references to them.

This problem manifested, for instance, when looking up events by 'mark'
(i.e. ":5"), for which evalAtomMark() calls getAttr() to retrieve "mark"
and "committish" fields from events.