eventToIndex: optimize for mark-based lookups
The Python version of this function attempts to do a fast lookup when
handed an event which has a 'mark' property, however, this functionality
was not carried over when ported to Go, so the Go code only does a slow
linear search over all events. Address this shortcoming by employing the
same fast map-based lookup as used by the Python version.