evalPathsetRegex: fix call to Commit.paths() always returning empty set
Commit.paths() returns the list of paths touched by the commit, filtered
by the 'pathtype' argument which specifies one or more fileops. If
'pathtype' is nil, it defaults to the set ("M", "D", "R", "C", "N").

The 'flags' argument to evalPathsetRegex() is composed of possible
fileop operations plus some custom flags (such as "a" and "c") which it
strips out before passing to Commit.paths(). However, 'flags' is never
nil; it may be empty (before or after stripping the custom flags) but
empty is not the same as nil, and FileOp.paths(), called by
Commits.paths(), does not recognize an empty 'pathtype' as an alias for
nil. Consequently, Commit.paths() would just return an empty set when no
fileops were specified explicitly. This problem manifested as either
[/.../] matching too few events because Commit.paths() was
unconditionally returning an empty set, or [/.../a] matching too many
paths because the "a" flag selects events for which _all_ paths are
matched by /.../, even if "all paths" is the empty set (and
Commit.paths() unconditionally returned the empty set).

One possible fix would be to upgrade FileOp.paths() to interpret both
nil and an empty 'pathtype' the same, however, all other callers of
FileOp.paths() which want default matching behavior consistently pass in
nil, so make evalPathsetRegex() do the same if 'flags' is empty.